backend/requirements.txt
alembic==1.9.2
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
black==22.12.0
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
coverage==7.2.0
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
psycopg2==2.9.5
pydantic==1.10.4
PyJWT==2.6.0
pytest==7.2.1
pytest-cov==4.0.0
python-dotenv==0.21.1
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
